pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sdc-agents"
version = "4.0.0"
description = "Purpose-scoped ADK agents for SDC4 data operations"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Timothy W. Cook", email = "tim@semanticdatacharter.org" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-adk>=1.25",
    "pydantic>=2",
    "pyyaml>=6",
    "httpx>=0.27",
    "sqlalchemy>=2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5",
    "black>=24",
    "ruff>=0.5",
    "aiosqlite>=0.20",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sdc_agents"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 99
target-version = ["py311"]
